+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- support for enhanced gender options from Shopgate
+### Updated
+- uses Shopgate Cart Integration SDK 2.9.91
 
 ## [2.9.32] - 2022-11-24
 ### Fixed